refactor: web, console 패키지 분리
