+++ src/main/java/roomescape/RoomescapeApplication.java
@@ -3,7 +3,7 @@ package roomescape;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.ConfigurableApplicationContext;
-import roomescape.controller.ConsoleMainController;
+import roomescape.console.controller.ConsoleMainController;
 
 @SpringBootApplication
 public class RoomescapeApplication {
