+++ src/main/java/roomescape/console/controller/CommandConverter.java
@@ -1,0 +1,11 @@
+package roomescape.console.controller;
+
+import java.util.Arrays;
+import java.util.List;
+
+public class CommandConverter {
+
+    public static List<String> convertPostReservation(final String read) {
+        return Arrays.stream(read.split("/")).map(String::trim).toList();
+    }
+}
