--- src/main/java/roomescape/controller/CommandConverter.java
@@ -1,11 +1,0 @@
-package roomescape.controller;
-
-import java.util.Arrays;
-import java.util.List;
-
-public class CommandConverter {
-
-    public static List<String> convertPostReservation(final String read) {
-        return Arrays.stream(read.split("/")).map(String::trim).toList();
-    }
-}
