--- src/main/java/roomescape/dto/ReservationRequest.java
@@ -1,13 +1,0 @@
-package roomescape.dto;
-
-import java.time.LocalDate;
-import java.util.Objects;
-
-public record ReservationRequest(String name, LocalDate date, Long timeId) {
-
-    public ReservationRequest {
-        Objects.requireNonNull(name);
-        Objects.requireNonNull(date);
-        Objects.requireNonNull(timeId);
-    }
-}
