+++ src/main/java/roomescape/general/dto/ReservationRequest.java
@@ -1,0 +1,13 @@
+package roomescape.general.dto;
+
+import java.time.LocalDate;
+import java.util.Objects;
+
+public record ReservationRequest(String name, LocalDate date, Long timeId) {
+
+    public ReservationRequest {
+        Objects.requireNonNull(name);
+        Objects.requireNonNull(date);
+        Objects.requireNonNull(timeId);
+    }
+}
