+++ src/main/java/roomescape/general/dto/ReservationTimeRequest.java
@@ -1,0 +1,11 @@
+package roomescape.general.dto;
+
+import java.time.LocalTime;
+import java.util.Objects;
+
+public record ReservationTimeRequest(LocalTime startAt) {
+
+    public ReservationTimeRequest {
+        Objects.requireNonNull(startAt);
+    }
+}
