+++ src/test/java/roomescape/db/ReservationTimeDaoH2ImplTest.java
@@ -7,7 +7,8 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
-import roomescape.domain.ReservationTime;
+import roomescape.general.domain.ReservationTime;
+import roomescape.web.db.ReservationTimeDaoH2Impl;
 
 @JdbcTest
 class ReservationTimeDaoH2ImplTest {
