+++ src/test/java/roomescape/domain/ReservationTest.java
@@ -3,6 +3,7 @@ package roomescape.domain;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import roomescape.general.domain.Reservation;
 
 class ReservationTest {
     @DisplayName("값이 null이면 예외가 발생한다")
