+++ src/test/java/roomescape/domain/ReservationTimeTest.java
@@ -1,10 +1,9 @@
 package roomescape.domain;
 
-import static org.junit.jupiter.api.Assertions.*;
-
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import roomescape.general.domain.ReservationTime;
 
 class ReservationTimeTest {
     @DisplayName("값이 null이면 예외가 발생한다")
